refactor(types): separa tipos por dominio e consolida compatibilidade
