+++ front-end/src/features/providers/types/provider.types.ts
@@ -1,0 +1,10 @@
+export interface Provider {
+  id: string;
+  displayId?: number;
+  name: string;
+  type: string;
+  status: string;
+  contact: string;
+  email: string;
+  address: string;
+}
