+++ front-end/src/shared/types/common.types.ts
@@ -1,0 +1,26 @@
+export type NavItem =
+  | 'dashboard'
+  | 'platformTenants'
+  | 'tenantProfile'
+  | 'revenues'
+  | 'payables'
+  | 'expenses'
+  | 'vehicles'
+  | 'suppliers'
+  | 'companies'
+  | 'contracts'
+  | 'freights'
+  | 'reports'
+  | 'settings'
+  | 'support';
+
+export interface UserProfile {
+  uid: string;
+  email: string;
+  role: 'dev' | 'owner' | 'admin' | 'financial' | 'operational' | 'driver' | 'viewer';
+  name?: string;
+  tenantId: string;
+  tenantName: string;
+  tenantSlug: string;
+  tenantLogoUrl?: string;
+}
